pyproject.toml
[tool.poetry]
name = "excel2anki"
version = "0.1.0"
description = ""
authors = ["Martin Buchleitner <martin@nitram.at>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12, <3.13"
genanki = "^0.13.1"
pandas = "^2.2.3"
openpyxl = "^3.1.5"
argparse = "^1.4.0"
openai = "^1.51.2"
flask = "^3.0.3"
requests = "^2.32.3"
pillow = "^11.0.0"
pytesseract = "^0.3.13"
opencv-python = "^4.10.0.84"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
